spatial/requirements.txt
torchvision>=0.15.0
transformers>=4.30.0
timm>=0.9.0
Pillow>=9.5.0
numpy>=1.24.0
pyyaml>=6.0
tensorboard>=2.13.0
tqdm>=4.65.0

